rely on fps for fastapi plugins and cli refactor plugins to use fps update CI to install contents plugin add flake8 and pre-commit configs
