+++ .pre-commit-config.yaml
@@ -1,0 +1,32 @@
+repos:
+  - repo: https://github.com/pycqa/isort
+    rev: 5.7.0
+    hooks:
+      - id: isort
+  - repo: https://github.com/psf/black
+    rev: 20.8b1
+    hooks:
+      - id: black
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.8.4
+    hooks:
+      - id: flake8
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v0.902
+    hooks:
+      - id: mypy
+        name: mypy-jupyverse
+        files: ^jupyverse/
+        args: [--show-error-codes]
+      - id: mypy
+        name: mypy-fps-contents
+        files: ^plugins/fps_contents/
+        args: [--show-error-codes]
+      - id: mypy
+        name: mypy-fps-kernels
+        files: ^plugins/fps_kernels/
+        args: [--show-error-codes]
+      - id: mypy
+        name: mypy-fps-retrolab
+        files: ^plugins/fps_retrolab/
+        args: [--show-error-codes]
