+++ jupyverse/__init__.py
@@ -1,4 +1,1 @@
-from .app import Jupyverse  # noqa
-from .router import JAPIRouter  # noqa
-
 from ._version import __version__  # noqa
