--- jupyverse/routers/contents/__init__.py
@@ -1,1 +1,0 @@
-from .routes import init  # noqa
