--- jupyverse/routers/jupyter_lab/__init__.py
@@ -1,1 +1,0 @@
-from .routes import init  # noqa
