+++ plugins/jupyterlab/setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup  # type: ignore
+
+setup(
+    name="fps_jupyterlab",
+    install_requires=["fps", "jupyterlab", "fps-contents", "fps-kernels"],
+    entry_points={
+        "fps_router": ["fps-jupyterlab = fps_jupyterlab.routes"],
+        "fps_config": ["fps-jupyterlab = fps_jupyterlab.config"],
+    },
+)
